+++ src/rules/rule-non-nested-empty-line-before/index.js
@@ -29,6 +29,7 @@ export default function (expectation, options) {
       actual: options,
       possible: {
         ignore: ["after-comment"],
+        except: ["after-single-line-comment"],
       },
       optional: true,
     })
@@ -57,7 +58,7 @@ export function checkRuleEmptyLineBefore({ rule, expectation, options, result, m
   if (optionsHaveIgnored(options, "after-comment")
     && rule.prev() && rule.prev().type === "comment") { return }
 
-  // Ignore if the exceptation is for multiple and the rule is single-line
+  // Ignore if the expectation is for multiple and the rule is single-line
   if (expectation.indexOf("multi-line") !== -1
     && isSingleLineString(rule.toString())) { return }
 
@@ -67,12 +68,22 @@ export function checkRuleEmptyLineBefore({ rule, expectation, options, result, m
     expectEmptyLineBefore = !expectEmptyLineBefore
   }
 
+  // Optionally reverse the expectation for single line comments
+  if (
+    optionsHaveException(options, "after-single-line-comment")
+    && rule.prev()
+    && rule.prev().type === "comment"
+    && isSingleLineString(rule.prev().toString())
+  ) {
+    expectEmptyLineBefore = !expectEmptyLineBefore
+  }
+
   const before = rule.raw("before")
   const emptyLineBefore = before && before.indexOf("\n\n") !== -1
     || before.indexOf("\r\n\r\n") !== -1
     || before.indexOf("\n\r\n") !== -1
 
-  // Return if the exceptation is met
+  // Return if the expectation is met
   if (expectEmptyLineBefore === emptyLineBefore) { return }
 
   const message = expectEmptyLineBefore ? messages.expected : messages.rejected